feat: I Hid the cases in two dimensions
Using the functions ind2sub2D and sub2ind2D directly can be slightly
more efficient but gives the user more power to mess things up. So, I
decided to use these two functions only internally in the other two
functions that converting indexes.
